Improved Debugability
Add variables that show the value saved to registers[pc->out]. When stepping through an executing program in the debugger this makes it easier to track the values coming out of each operation.
Also adds a few quality of life instructions that turned out to be useful when writing bigger programs.